[Psiphon] Resets PRNG for marshalRandomized
Resets the PRNG used in marshalRandomized since it is called more than
once for the same Client Hello with and without PSK binders.

Caching marshalRandomizedNoECH might be possible; appending the
PSK binders when present. However this would conflict with the
behaviour of marshalMsg in upstream crypto/tls, violating the assumptions
made by the callers.

https://cs.opensource.google/go/go/+/f0d6ddfac03c1c3048f20d595fd50352c79e1849